refactor: 다나와 크롤러 계층 분리 및 메서드 추출
