+++ backend/src/main/java/codebadger/virtual_launch/domain/crawling/application/CompetitorCrawlingService.java
@@ -1,0 +1,12 @@
+package codebadger.virtual_launch.domain.crawling.application;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class CompetitorCrawlingService {
+
+}
