+++ backend/src/main/java/codebadger/virtual_launch/domain/crawling/infrastructure/DanawaSpecCrawler.java
@@ -1,0 +1,4 @@
+package codebadger.virtual_launch.domain.crawling.infrastructure;
+
+public class DanawaSpecCrawler {
+}
